requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.14.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
